+++ app/src/routes/editor/EditPanel/ManualTab/compound/DisplayInput.tsx
