+++ app/src/routes/editor/EditPanel/ManualTab/compound/NestedInputs.tsx
