+++ app/src/routes/editor/EditPanel/ManualTab/index.tsx
