+++ app/src/routes/editor/EditPanel/ManualTab/single/ColorInput/index.tsx
