+++ app/src/routes/editor/EditPanel/ManualTab/single/SelectInput.tsx
