+++ app/src/routes/editor/EditPanel/ManualTab/single/TagDetails.tsx
