+++ app/src/routes/editor/EditPanel/ManualTab/single/TextInput.tsx
